requirements.txt
chromadb==0.4.23
ctransformers==0.2.27
InstructorEmbedding==1.0.1
langchain==0.1.9
langchain-community==0.0.22
llama-cpp-python==0.2.20
librosa==0.10.1
pypdfium2==4.27.0
pyyaml==6.0.1
sentence-transformers==2.2.2
streamlit==1.31.1
streamlit-mic-recorder==0.0.4
transformers==4.38.1
sentence-transformers==2.2.2


# aiohttp==3.9.3
# aiosignal==1.3.1
# altair==5.2.0
# annotated-types==0.6.0
# anyio==4.2.0
# asgiref==3.7.2
# async-timeout==4.0.3
# attrs==23.2.0
# audioread==3.0.1
# backoff==2.2.1
# bcrypt==4.1.2
# blinker==1.7.0
# build==1.0.3
# cachetools==5.3.2
# certifi==2022.12.7
# cffi==1.16.0
# charset-normalizer==2.1.1
# chroma-hnswlib==0.7.3
# chromadb==0.4.22
# click==8.1.7
# coloredlogs==15.0.1
# ctransformers==0.2.27
# dataclasses-json==0.6.4
# decorator==5.1.1
# Deprecated==1.2.14
# diskcache==5.6.3
# exceptiongroup==1.2.0
# fastapi==0.109.2
# filelock==3.9.0
# flatbuffers==23.5.26
# frozenlist==1.4.1
# fsspec==2024.2.0
# gitdb==4.0.11
# GitPython==3.1.41
# google-auth==2.27.0
# googleapis-common-protos==1.62.0
# greenlet==3.0.3
# grpcio==1.60.1
# h11==0.14.0
# httptools==0.6.1
# huggingface-hub==0.20.3
# humanfriendly==10.0
# idna==3.4
# importlib-metadata==6.11.0
# importlib-resources==6.1.1
# InstructorEmbedding==1.0.1
# Jinja2==3.1.2
# joblib==1.3.2
# jsonpatch==1.33
# jsonpointer==2.4
# jsonschema==4.21.1
# jsonschema-specifications==2023.12.1
# kubernetes==29.0.0
# langchain==0.1.6
# langchain-community==0.0.19
# langchain-core==0.1.22
# langsmith==0.0.87
# lazy_loader==0.3
# librosa==0.10.1
# llama_cpp_python==0.2.39
# llvmlite==0.42.0
# markdown-it-py==3.0.0
# MarkupSafe==2.1.3
# marshmallow==3.20.2
# mdurl==0.1.2
# mmh3==4.1.0
# monotonic==1.6
# mpmath==1.3.0
# msgpack==1.0.7
# multidict==6.0.5
# mypy-extensions==1.0.0
# networkx==3.2.1
# nltk==3.8.1
# numba==0.59.0
# numpy==1.26.3
# oauthlib==3.2.2
# onnxruntime==1.17.0
# opentelemetry-api==1.22.0
# opentelemetry-exporter-otlp-proto-common==1.22.0
# opentelemetry-exporter-otlp-proto-grpc==1.22.0
# opentelemetry-instrumentation==0.43b0
# opentelemetry-instrumentation-asgi==0.43b0
# opentelemetry-instrumentation-fastapi==0.43b0
# opentelemetry-proto==1.22.0
# opentelemetry-sdk==1.22.0
# opentelemetry-semantic-conventions==0.43b0
# opentelemetry-util-http==0.43b0
# overrides==7.7.0
# packaging==23.2
# pandas==2.2.0
# pillow==10.2.0
# platformdirs==4.2.0
# pooch==1.8.0
# posthog==3.4.0
# protobuf==4.25.2
# pulsar-client==3.4.0
# py-cpuinfo==9.0.0
# pyarrow==15.0.0
# pyasn1==0.5.1
# pyasn1-modules==0.3.0
# pycparser==2.21
# pydantic==2.6.1
# pydantic_core==2.16.2
# pydeck==0.8.1b0
# Pygments==2.17.2
# pypdfium2==4.26.0
# PyPika==0.48.9
# pyproject_hooks==1.0.0
# python-dateutil==2.8.2
# python-dotenv==1.0.1
# pytz==2024.1
# PyYAML==6.0.1
# referencing==0.33.0
# regex==2023.12.25
# requests==2.28.1
# requests-oauthlib==1.3.1
# rich==13.7.0
# rpds-py==0.17.1
# rsa==4.9
# safetensors==0.4.2
# scikit-learn==1.4.0
# scipy==1.12.0
# sentence-transformers==2.2.2
# sentencepiece==0.1.99
# six==1.16.0
# smmap==5.0.1
# sniffio==1.3.0
# soundfile==0.12.1
# soxr==0.3.7
# SpeechRecognition==3.10.1
# SQLAlchemy==2.0.25
# starlette==0.36.3
# streamlit==1.31.0
# streamlit-mic-recorder==0.0.4
# sympy==1.12
# tenacity==8.2.3
# threadpoolctl==3.2.0
# tokenizers==0.15.1
# toml==0.10.2
# tomli==2.0.1
# toolz==0.12.1
# torch==2.2.0+cpu
# torchaudio==2.2.0+cpu
# torchvision==0.17.0+cpu
# tornado==6.4
# tqdm==4.66.1
# transformers==4.37.2
# typer==0.9.0
# typing-inspect==0.9.0
# typing_extensions==4.8.0
# tzdata==2023.4
# tzlocal==5.2
# urllib3==1.26.13
# uvicorn==0.27.0.post1
# uvloop==0.19.0
# validators==0.22.0
# watchdog==4.0.0
# watchfiles==0.21.0
# websocket-client==1.7.0
# websockets==12.0
# wrapt==1.16.0
# yarl==1.9.4
# zipp==3.17.0
